+++ src/oop/part2/Bicycle.java
@@ -1,24 +1,9 @@
 package oop.part2;
 
-public class Bicycle implements RepairTransport {
-
-    private String modelName;
-    private int wheelsCount;
+public class Bicycle extends GeneralPropertiesTransport {
 
     public Bicycle(String modelName, int wheelsCount) {
-        this.modelName = modelName;
-        this.wheelsCount = wheelsCount;
+        super(modelName, wheelsCount);
     }
 
-    public void updateTyre() {
-        System.out.println("Меняем покрышку");
-    }
-
-    @Override
-    public void service() {
-        System.out.println("Обслуживаем " + modelName);
-        for (int i = 0; i < wheelsCount; i++) {
-            updateTyre();
-        }
-    }
 }
