+++ src/oop/part2/Car.java
@@ -1,18 +1,9 @@
 package oop.part2;
 
-public class Car implements RepairTransport {
-
-    private String modelName;
-    private int wheelsCount;
-
+public class Car extends GeneralPropertiesTransport {
 
     public Car(String modelName, int wheelsCount) {
-        this.modelName = modelName;
-        this.wheelsCount = wheelsCount;
-    }
-
-    public void updateTyre() {
-        System.out.println("Меняем покрышку");
+        super(modelName, wheelsCount);
     }
 
     public void checkEngine() {
@@ -21,10 +12,6 @@ public class Car implements RepairTransport {
 
     @Override
     public void service() {
-        System.out.println("Обслуживаем " + modelName);
-        for (int i = 0; i < wheelsCount; i++) {
-            updateTyre();
-        }
         checkEngine();
     }
 }
