+++ src/oop/part2/GeneralPropertiesTransport.java
@@ -1,0 +1,22 @@
+package oop.part2;
+
+public abstract class GeneralPropertiesTransport implements RepairTransport {
+    private String modelName;
+    private int wheelsCount;
+
+    public GeneralPropertiesTransport(String modelName, int wheelsCount) {
+        this.modelName = modelName;
+        this.wheelsCount = wheelsCount;
+    }
+    public void updateTyre() {
+        System.out.println("Меняем покрышку");
+    }
+    @Override
+    public void service() {
+        System.out.println("Обслуживаем " + modelName);
+        for (int i = 0; i < wheelsCount; i++) {
+            updateTyre();
+        }
+    }
+
+}
