+++ src/oop/part2/Main.java
@@ -11,6 +11,7 @@ public class Main {
         Bicycle bicycle = new Bicycle("bicycle1", 2);
         Bicycle bicycle2 = new Bicycle("bicycle2",2);
 
+
         serviceTransport(car);
         serviceTransport(car2);
         serviceTransport(truck);
