+++ src/oop/part2/Truck.java
@@ -1,18 +1,11 @@
 package oop.part2;
 
-public class Truck implements RepairTransport {
-
-    private String modelName;
-    private int wheelsCount;
+public class Truck extends GeneralPropertiesTransport {
 
     public Truck(String modelName, int wheelsCount) {
-        this.modelName = modelName;
-        this.wheelsCount = wheelsCount;
+        super(modelName, wheelsCount);
     }
 
-    public void updateTyre() {
-        System.out.println("Меняем покрышку");
-    }
 
     public void checkEngine() {
         System.out.println("Проверяем двигатель");
@@ -23,10 +16,6 @@ public class Truck implements RepairTransport {
     }
     @Override
     public void service() {
-        System.out.println("Обслуживаем " + modelName);
-        for (int i = 0; i < wheelsCount; i++) {
-            updateTyre();
-        }
         checkEngine();
         checkTrailer();
     }
